[Odoo-7634/7629][FIX][TIMESHEET_ODOO]

+++ timesheet_odoo/models/period_sige.py
@@ -22,6 +22,8 @@ class PeriodSige(models.Model):
         ("close","Close")
     ], "State", index=True, default="open")
 
+    confirmation = fields.Boolean(default=False)
+
     @api.constrains("start_of_period")
     def _unique_period_per_year(self):
         for record in self:
@@ -179,13 +181,16 @@ class PeriodSige(models.Model):
 
     @api.model
     def custom_delete_records(self):
-        selected_periods = self.env['period.sige'].browse(self.ids)
-        for period in selected_periods:
-            # Eliminar los registros relacionados en timesheet.sige
-            timesheets_to_delete = self.env['timesheet.sige'].search([('period_id', '=', period.id)])
-            timesheets_to_delete.unlink()
+        if self.confirmation:
+            selected_periods = self.env['period.sige'].browse(self.ids)
+            for period in selected_periods:
+                # Eliminar los registros relacionados en timesheet.sige
+                timesheets_to_delete = self.env['timesheet.sige'].search([('period_id', '=', period.id)])
+                timesheets_to_delete.unlink()
 
-            # Eliminar el periodo
-            period.unlink()
+                # Eliminar el periodo
+                period.unlink()
+        else:
+            raise UserError(_('You cannot permission to delete a period'))
